organize code and filenames
