+++ transforms/universal/ededup/dpk_ededup/__init__.py
@@ -1,4 +1,3 @@
 # SPDX-License-Identifier: Apache-2.0
 from .transform_base import *
-from .transform_python import *
-
+from .runtime import *
