+++ transforms/universal/ededup/dpk_ededup/runtime.py
@@ -19,43 +19,17 @@ from data_processing.runtime.pure_python import (
     DefaultPythonTransformRuntime,
     PythonTransformLauncher,
     PythonTransformRuntimeConfiguration,
+    Transform,
 )
 from data_processing.transform import TransformStatistics
 from data_processing.utils import ParamsUtils
 from dpk_ededup.transform_base import (
-    EdedupTransformBase,
+    EdedupTransform,
     EdedupTransformConfigurationBase,
     HashFilter,
     snapshot_directory_key,
     use_snapshot_key,
 )
-
-
-class EdedupTransform(EdedupTransformBase):
-    """
-    Implements dedup table transformer.
-    """
-
-    def __init__(self, config: dict):
-        """
-        Initialize based on the dictionary of configuration information.
-        The dictionary should contain the following:
-            doc_column - name of the doc column
-        """
-        super().__init__(config)
-        self.filter = config.get("filter", HashFilter({}))
-
-    def _process_cached_hashes(self, hd: dict[str, str]) -> list[str]:
-        """
-        check hashes uniqueness with the distributed cache of hashes
-        :param hd: dictionary of hash to document
-        :return: unique documents
-        """
-        unique_hashes = self.filter.get_unique(ha=list(hd.keys()))
-        unique = []
-        for uh in unique_hashes:
-            unique.append(hd[uh])
-        return unique
 
 
 class EdedupRuntime(DefaultPythonTransformRuntime):
@@ -144,27 +118,9 @@ class EdedupPythonTransformRuntimeConfiguration(PythonTransformRuntimeConfigurat
 
 
 # Class used by the notebooks to ingest binary files and create parquet files
-class Ededup:
+class Ededup(Transform):
     def __init__(self, **kwargs):
-        self.params = {}
-        for key in kwargs:
-            self.params[key] = kwargs[key]
-        # if input_folder and output_folder are specified, then assume it is represent data_local_config
-        try:
-            local_conf = {k: self.params[k] for k in ("input_folder", "output_folder")}
-            self.params["data_local_config"] = ParamsUtils.convert_to_ast(local_conf)
-            del self.params["input_folder"]
-            del self.params["output_folder"]
-        except:
-            pass
-
-    def transform(self):
-        sys.argv = ParamsUtils.dict_to_req(d=(self.params))
-        # create launcher
-        launcher = PythonTransformLauncher(EdedupPythonTransformRuntimeConfiguration())
-        # launch
-        return_code = launcher.launch()
-        return return_code
+        super().__init__(EdedupPythonTransformRuntimeConfiguration(), **kwargs)
 
 
 if __name__ == "__main__":
