+++ transforms/universal/ededup/dpk_ededup/transform_base.py
@@ -11,8 +11,8 @@
 # limitations under the License.
 ################################################################################
 
-from abc import abstractmethod
 import pickle
+from abc import abstractmethod
 from argparse import ArgumentParser, Namespace
 from typing import Any
 
@@ -40,7 +40,6 @@ doc_column_name_cli_param = f"{cli_prefix}{doc_column_name_key}"
 int_column_name_cli_param = f"{cli_prefix}{int_column_name_key}"
 use_snapshot_cli_param = f"{cli_prefix}{use_snapshot_key}"
 snapshot_directory_cli_param = f"{cli_prefix}{snapshot_directory_key}"
-
 
 
 class HashFilter:
@@ -178,7 +177,11 @@ class EdedupTransformBase(AbstractTableTransform):
         # Create output table
         out_table = table.filter(mask)
         # report statistics
-        stats = {"source_documents": table.num_rows, "result_documents": out_table.num_rows, "removed_documents": removed}
+        stats = {
+            "source_documents": table.num_rows,
+            "result_documents": out_table.num_rows,
+            "removed_documents": removed,
+        }
         return [out_table], stats
 
     @abstractmethod
@@ -244,3 +247,30 @@ class EdedupTransformConfigurationBase(TransformConfiguration):
         self.params = self.params | captured
         self.logger.info(f"exact dedup params are {self.params}")
         return True
+
+
+class EdedupTransform(EdedupTransformBase):
+    """
+    Implements dedup table transformer.
+    """
+
+    def __init__(self, config: dict):
+        """
+        Initialize based on the dictionary of configuration information.
+        The dictionary should contain the following:
+            doc_column - name of the doc column
+        """
+        super().__init__(config)
+        self.filter = config.get("filter", HashFilter({}))
+
+    def _process_cached_hashes(self, hd: dict[str, str]) -> list[str]:
+        """
+        check hashes uniqueness with the distributed cache of hashes
+        :param hd: dictionary of hash to document
+        :return: unique documents
+        """
+        unique_hashes = self.filter.get_unique(ha=list(hd.keys()))
+        unique = []
+        for uh in unique_hashes:
+            unique.append(hd[uh])
+        return unique
