+++ transforms/universal/ededup/test/test_ededup_python_incremental.py
@@ -17,13 +17,13 @@ from data_processing.runtime.pure_python import PythonTransformLauncher
 from data_processing.test_support.launch.transform_test import (
     AbstractTransformLauncherTest,
 )
+from dpk_ededup.runtime import EdedupPythonTransformRuntimeConfiguration
 from dpk_ededup.transform_base import (
     doc_column_name_cli_param,
     int_column_name_cli_param,
     snapshot_directory_cli_param,
-    use_snapshot_cli_param
+    use_snapshot_cli_param,
 )
-from dpk_ededup.transform_python import EdedupPythonTransformRuntimeConfiguration
 
 
 class TestPythonEdedupTransform(AbstractTransformLauncherTest):
