+++ transforms/universal/ededup/test/test_ededup_ray.py
@@ -17,14 +17,14 @@ from data_processing.test_support.launch.transform_test import (
     AbstractTransformLauncherTest,
 )
 from data_processing_ray.runtime.ray import RayTransformLauncher
-from dpk_ededup.ray.transform import (
+from dpk_ededup.ray.runtime import (
     EdedupRayTransformRuntimeConfiguration,
     hash_cpu_cli_params,
     num_hashes_cli_params,
 )
 from dpk_ededup.transform_base import (
     doc_column_name_cli_param,
-    int_column_name_cli_param
+    int_column_name_cli_param,
 )
 
 
@@ -43,7 +43,7 @@ class TestRayEdedupTransform(AbstractTransformLauncherTest):
             hash_cpu_cli_params: 0.5,
             num_hashes_cli_params: 2,
             doc_column_name_cli_param: "contents",
-            int_column_name_cli_param: "document_id"
+            int_column_name_cli_param: "document_id",
         }
         launcher = RayTransformLauncher(EdedupRayTransformRuntimeConfiguration())
         fixtures = [(launcher, config, basedir + "/input", basedir + "/expected")]
